Use `Homebrew::CLI::Parser`
This will improve argument parsing and `--help` output.
